+++ NEWS.md
@@ -3,6 +3,7 @@
 * development version
 * Added `new_module_template()` to create new module templates and `use_formods()` 
 to automatically add the files to a package. 
+* Updated app info in ASM to split up the uiele into diferent components. 
 
 # formods 0.1.1